requirements.txt
fastapi
uvicorn
stripe
pyodbc
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib
bcrypt